refact: move exception package
and refact kid personality package
