--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/common/exception/book/BookException.java
@@ -1,11 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.common.exception.book;
-
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.BaseException;
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.ExceptionType;
-
-public class BookException extends BaseException {
-
-    public BookException(ExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
