--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/common/exception/config/ExceptionResponse.java
@@ -1,6 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.common.exception.config;
-
-
-
-public record ExceptionResponse(String message) {}
-
