+++ src/main/java/com/ureca/idle/idleapi/idleoriginapi/implementation/book/BookException.java
@@ -1,0 +1,11 @@
+package com.ureca.idle.idleapi.idleoriginapi.implementation.book;
+
+import com.ureca.idle.idleexception.BaseException;
+import com.ureca.idle.idleexception.ExceptionType;
+
+public class BookException extends BaseException {
+
+    public BookException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
