+++ src/main/java/com/ureca/idle/idleapi/idleoriginapi/implementation/kid/KidException.java
@@ -1,0 +1,14 @@
+package com.ureca.idle.idleapi.idleoriginapi.implementation.kid;
+
+import com.ureca.idle.idleexception.BaseException;
+import com.ureca.idle.idleexception.ExceptionType;
+
+
+
+public class KidException extends BaseException {
+
+    public KidException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
+
